refactor(headless): split Tooltip into per-file parts for tree-shaking
