+++ packages/headless/src/primitives/tooltip/parts.ts
@@ -1,0 +1,7 @@
+export { type TooltipProps, TooltipRoot as Root } from './tooltip-root';
+export { type TooltipTriggerProps, TooltipTrigger as Trigger } from './tooltip-trigger';
+export { type TooltipPortalProps, TooltipPortal as Portal } from './tooltip-portal';
+export { type TooltipPositionerProps, TooltipPositioner as Positioner } from './tooltip-positioner';
+export { type TooltipPopupProps, TooltipPopup as Popup } from './tooltip-popup';
+export { type TooltipArrowProps, TooltipArrow as Arrow } from './tooltip-arrow';
+export { type TooltipGroupProps, TooltipGroup as Group } from './tooltip-group';
